Created a branch
I seemed to do the same as in homework solution but wanted to also try if I could achieve something similar with one row occupied with book card and comment reveal would come to side when clicked. Works but in a way that click on book causes the comment to always show up on the container. How would it be possible to have comment section showing up next to book individually?